Extract helper with logic to expand the search term for a query
