+++ app/servers/ServersRepository.server.ts
@@ -1,5 +1,6 @@
 import type { EntityManager, FilterQuery } from '@mikro-orm/core';
 import { BaseEntityRepository } from '../db/BaseEntityRepository';
+import { expandSearchTerm } from '../db/utils.server';
 import { Server } from '../entities/Server';
 
 export type FindServersOptions = {
@@ -16,38 +17,13 @@ export class ServersRepository extends BaseEntityRepository<Server> {
   }
 
   findByUserId(userId: string, { searchTerm, populateUsers = false }: FindServersOptions = {}): Promise<Server[]> {
-    return this.find(this.buildListServersWhere(userId, searchTerm), {
-      orderBy: { name: 'ASC' },
-      populate: populateUsers ? ['users'] : undefined,
-    });
-  }
-
-  private buildListServersWhere(userId: string, searchTerm?: string): FilterQuery<Server> {
     const baseFilter: FilterQuery<Server> = {
       users: { id: userId },
     };
-
-    if (!searchTerm) {
-      return baseFilter;
-    }
-
-    const searchableFields: Array<keyof Server> = ['name', 'baseUrl'];
-    const lowerSearchTerm = searchTerm.toLowerCase();
-
-    baseFilter['$or'] = searchableFields.flatMap((field) => [
-      {
-        [field]: {
-          $like: `%${lowerSearchTerm}%`,
-        },
-      },
-      {
-        [field]: {
-          $like: `%${searchTerm}%`,
-        },
-      },
-    ]);
-
-    return baseFilter;
+    return this.find(expandSearchTerm<Server>(searchTerm, { searchableFields: ['name', 'baseUrl'], baseFilter }), {
+      orderBy: { name: 'ASC' },
+      populate: populateUsers ? ['users'] : undefined,
+    });
   }
 }
